+++ chainer/links/__init__.py
@@ -29,6 +29,7 @@ from chainer.links.loss import negative_sampling  # NOQA
 from chainer.links.model import classifier  # NOQA
 from chainer.links.normalization import batch_normalization  # NOQA
 from chainer.links.normalization import layer_normalization  # NOQA
+from chainer.links.theano import theano_function  # NOQA
 
 
 # import class and function
@@ -65,3 +66,4 @@ from chainer.links.model.vision.resnet import ResNet50Layers  # NOQA
 from chainer.links.model.vision.vgg import VGG16Layers  # NOQA
 from chainer.links.normalization.batch_normalization import BatchNormalization  # NOQA
 from chainer.links.normalization.layer_normalization import LayerNormalization  # NOQA
+from chainer.links.theano.theano_function import TheanoFunction  # NOQA
